Se modifico el reporte de las lista de luces. Se modificaron algunas migraciones.

+++ database/migrations/2018_04_24_162034_create_period_table.php
@@ -15,7 +15,7 @@ class CreatePeriodTable extends Migration
     {
         Schema::create('period', function (Blueprint $table) {
             $table->increments('id')->unsigned();
-            $table->integer('time');
+            $table->decimal('time', 10, 2);
             $table->integer('flash_group');
             $table->timestamps();
             $table->string('created_by', 100);
